Finish Set 1 Challenge 6
Finally finished this...

There's a bug in the base64 decode function that I implemented...need to
figure out what's happening with that.

I added a Chi Squared scoring function to the operations package.

I changed some of the return and input types for existing functions when
I was trying to debug this...no sense in changing them back now.